# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": " NodeJS API project to demonstrate Swagger ",
        "version": "1.0.0",
        "title": "E-commers Project API"
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:7065",
    "basePath": "/api/v1",
    "paths": {
...
        "/create/order/add-order/{user_Id}": {
            "post": {
                "summary": "Save the Order",
                "description": "Create the Order ",
                "produces": [
                    "application/json"
                ],
                "requestBody": true,
                "parameters": [ 
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Order body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "productId": {
                                    "type": "string",
                                    "example": "656597a9a08d38ee8ada8f23"
                                },
                                "quantity": {
                                    "type": "number"
                                },
                                "shipping_address": {
                                    "type": "string",
                                    "example": "logo"
                                }
                            }
                        }
                    } ,   
                    {
                        "name": "user_Id",
                        "in": "path",
                        "description": "user_Id",
                        "required": true,
                        "type": "string"
                    },
                    
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/orderResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/order/get-all-order": {
            "get": {
                "summary": "Get All  Order",
                "description": "Get All  Order !!!",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },

                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                    
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/orderResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/order/get-order/{order_Id}/{user_Id}": {
            "get": {
                "summary": "Get the Order",
                "description": "Get the Order",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "order_Id",
                        "in": "path",
                        "description": "order_Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_Id",
                        "in": "path",
                        "description": "user_Id",
                        "required": true,
                        "type": "string"
                    },
                    
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/orderResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }, 
        "/order/get-total_Product_Sale-order": {
            "get": {
                "summary": "Get the Total Product Sale Order",
                "description": "Get the Total Product Sale Order",
                "produces": [
                    "application/json"
                ],
                "parameters": [ 
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/total_Product_Sale-order"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/order/get-total_User_Sale/{user_Id}": {
            "get": {
                "summary": "Get Total User Sale",
                "description": "Get Total User Sale",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                   
                    {
                        "name": "user_Id",
                        "in": "path",
                        "description": "user_Id",
                        "required": true,
                        "type": "string"
                    },
                    
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/total_User_Sale"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/order/update-order/{order_Id}": {
            "put": {
                "summary": "Update Order Status",
                "description": "Update Order ",
                "produces": [
                    "application/json"
                ], 
                "consumes": ["application/json"],
                "requestBody": true,
                "parameters": [{
                        "name": "updateData",
                        "in": "body",
                        "description": " body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example" : "delivered"
                                }
                            }                        
                        }
                    },
                    {
                        "name": "order_Id",
                        "in": "path",
                        "description": "order_Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/update-order_user"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/order/update-order_user/{user_Id}/{order_Id}": {
            "put": {
                "summary": "Update Order Ststus By User",
                "description": "Update Order ",
                "produces": [
                    "application/json"
                ], 
                "consumes": ["application/json"],
                "requestBody": true,
                "parameters": [{
                        "name": "user_Id",
                        "in": "path",
                        "description": "user_Id",
                        "required": true,
                        "type": "string"
                    },{
                        "name": "order_Id",
                        "in": "path",
                        "description": "order_Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/update-order_user"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/order/delete-order/{order_Id}": {
            "delete": {
                "summary": "Delete the Order",
                "description": "Delete the Order",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "order_Id",
                        "in": "path",
                        "description": "order_Id",
                        "required": true,
                        "type": "string"
                    },
                    
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/deleteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "orderResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "true"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "data": {
                    "type": "object",                  
                        "properties": {
                            "userId": {
                                "type": "string",
                                "example": "mobile"
                            },
                            "productId": {
                                "type": "array",
                                "items": {
                                  "type": "string",
                                  "example": "658be1b47d2f886a923c69b5"
                                }
                              },
                              "quantity": {
                                "type": "number",
                                "example": 3
                              },
                              "shipping_address": {
                                "type": "string",
                                "example": "lucknow"
                              },
                              "status": {
                                "type": "string",
                                "example": "pending"
                              },
                              "isDeleted": {
                                "type": "boolean",
                                "example": false
                              },
                              "_id": {
                                "type": "string",
                                "example": "658e45d7911973e54d3cfb73"
                              },
                            "createdAt": { "type": "string", "format": "date-time", "example": "2023-11-30T07:52:22.910Z" },
                            "updatedAt": { "type": "string", "format": "date-time", "example": "2023-11-30T07:52:22.910Z" }
                        }
                
                    
                }
            }
        },
        "deleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted successfully"
                },
                "Session": {
                    "type": "object",
                    "properties": {
                        "acknowledged": {
                            "type": "string",
                            "example": "true"
                        },
                        "deletedCount": {
                            "type": "number",
                            "example": "2"
                        }
                    }
                }
            }
        },
        "total_User_Sale": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "example": "true"
              },
              "message": {
                "type": "string",
                "example": "Success"
              },
              "data": {
                "type": "object",
                "properties": {
                  "user_Id": {
                    "type": "string",
                    "example": "658c533941d730477b31529b"
                  },
                  "total_Sale": {
                    "type": "number",
                    "example": 53
                  },
                  "total_Cancelled": {
                    "type": "number",
                    "example": 1
                  },
                  "total_Delivered": {
                    "type": "number",
                    "example": 0
                  },
                  "Product_bar": {
                    "type": "object",
                    "properties": {
                      "Toy": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Product_Order": {
                                    "type": "number",
                                    "example": 2
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "Bottle": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Product_Order": {
                                    "type": "number",
                                    "example": 8
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "mobile": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Product_Order": {
                                    "type": "number",
                                    "example": 1
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "mobiole ty": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Product_Order": {
                                    "type": "number",
                                    "example": 5
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "status_bar": {
                    "type": "object",
                    "properties": {
                      "pending": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Order_Status": {
                                    "type": "number",
                                    "example": 52
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "cancelled": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Order_Status": {
                                    "type": "number",
                                    "example": 1
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
        },
        "total_Product_Sale-order": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "example": "true"
              },
              "message": {
                "type": "string",
                "example": "Success"
              },
              "data": {
                "type": "object",
                "properties": {
                  "total_Sale": {
                    "type": "number",
                    "example": 53
                  },
                  "Product_bar": {
                    "type": "object",
                    "properties": {
                      "Toy": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Product_Order": {
                                    "type": "number",
                                    "example": 2
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "Bottle": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Product_Order": {
                                    "type": "number",
                                    "example": 8
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "mobile": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Product_Order": {
                                    "type": "number",
                                    "example": 1
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "mobiole ty": {
                        "type": "object",
                        "properties": {
                          "2023": {
                            "type": "object",
                            "properties": {
                              "dec": {
                                "type": "object",
                                "properties": {
                                  "total_Product_Order": {
                                    "type": "number",
                                    "example": 5
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                 
                }
              }
            }
        },
        "update-order_user": {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "example": true
              },
              "message": {
                "type": "string",
                "example": "Order canceled successfully and email sent"
              },
              "data": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "example": "658d269add8884cd6a415953"
                  },
                  "userId": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "658c533941d730477b31529b"
                      },
                      "userName": {
                        "type": "string",
                        "example": "Sagarsss"
                      },
                      "email": {
                        "type": "string",
                        "example": "prachisingh281599@gmail.com"
                      },
                      "address": {
                        "type": "string",
                        "example": "13/17 colonel ganj kanpur"
                      }
                    }
                  },
                  "productId": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "658be1b47d2f886a923c69b5"
                        },
                        "product_Name": {
                          "type": "string",
                          "example": "Bottle"
                        },
                        "category": {
                          "type": "string",
                          "example": "milton"
                        },
                        "price": {
                          "type": "number",
                          "example": 800
                        },
                        "image_URL": {
                          "type": "string",
                          "example": "bbmobile"
                        }
                      }
                    }
                  },
                  "quantity": {
                    "type": "number",
                    "example": 78
                  },
                  "shipping_address": {
                    "type": "string",
                    "example": "j"
                  },
                  "status": {
                    "type": "string",
                    "example": "cancelled"
                  },
                  "isDeleted": {
                    "type": "boolean",
                    "example": false
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-12-28T07:41:14.562Z"
                  },
                  "updatedAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-12-28T07:45:08.643Z"
                  },
                  "__v": {
                    "type": "number",
                    "example": 0
                  }
                }
              }
            }
          },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "Number",
                    "example": "400"
                   
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}